requirements.txt
python-decouple
SQLAlchemy
uvicorn
fastapi
asyncpg
pytest
pytest-asyncio